+++ src/Submissions/Api/Views/Shared/_FeedbackMessage.cshtml
@@ -1,6 +1,6 @@
 ﻿@using Biobanks.Submissions.Api.Controllers
 @using Biobanks.Submissions.Api.Utilities;
-@using static Biobanks.Submissions.Api.Controllers.Submissions.ApplicationBaseController
+@using static Biobanks.Submissions.Api.Controllers.Directory.ApplicationBaseController
 @{
     var model = (FeedbackMessage)TempData[ViewConstants.FeedbackMessageKey];
 }
