+++ src/Submissions/Api/appsettings.json
@@ -1,6 +1,6 @@
 {
   "ConnectionStrings": {
-    "Default": "Server=(localdb)\\MSSQLLocalDB;Database=Biobanks;Trusted_Connection=True;MultipleActiveResultSets=true",
+    "Default": "Server=localhost;Database=Biobanks;MultipleActiveResultSets=true;User ID=SA;Password=Password#1;",
     "Omop": "Server=atlas-maps.postgres.database.azure.com;Database=omop",
     "AzureStorage": "UseDevelopmentStorage=true;"
   },
